refactor: 댓글 섹션 simplify 정리 (#49)
- CommentSection → EpigramDetailList 위젯 폴더 개명
- CommentItem 별도 파일로 분리
- FlatList ListHeader/Empty/Footer를 element로 전달해 TextInput 재마운트 방지
- ItemSeparator 모듈 레벨 추출, renderItem useCallback
- useCommentDelete 내부 confirm 제거 (action sheet 의존), guard clause 적용
- useCommentCreate hasError를 mutation isError에서 파생
- WriterAvatar 인터페이스 좁혀 User에서도 재사용
- PrivacyToggle 공용 컴포넌트 추출, CommentForm/CommentEditForm에서 공유
