+++ src/shared/ui/index.ts
@@ -2,3 +2,4 @@ export { Button } from "./Button";
 export type { ButtonProps } from "./Button";
 export { Input } from "./Input";
 export type { InputProps } from "./Input";
+export { PrivacyToggle } from "./PrivacyToggle";
