--- src/widgets/comment-section/index.ts
@@ -1,1 +1,0 @@
-export { CommentSection } from "./ui/CommentSection";
