+++ src/widgets/epigram-detail-list/index.ts
@@ -1,0 +1,1 @@
+export { EpigramDetailList } from "./ui/EpigramDetailList";
